backend/requirements.txt
# FastAPI & server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# PostGIS / Geo
geoalchemy2>=0.14.0

# Redis
redis>=5.0.0

# Celery
celery[redis]>=5.3.0

# LLM
openai>=1.12.0
langgraph>=0.0.20
langchain>=0.1.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Web Push
pywebpush>=1.14.0

# HTTP client
httpx>=0.26.0

# Dev / tooling
ruff>=0.2.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
